Move codes to error package
