+++ ERROR_SPECIFICATION.md
@@ -1,0 +1,1 @@
+# Error Module
