+++ core/error/codes.go
@@ -1,4 +1,4 @@
-package codes
+package error
 
 // A Code is an unsigned 32-bit error code as defined in the beego spec.
 type Code uint32
@@ -10,5 +10,5 @@ const (
 
 // CodeToStr is a map about Code and Code's message
 var CodeToStr = map[Code]string{
-	SessionSessionStartError : `"SESSION_MODULE_SESSION_START_ERROR"`,
+	SessionSessionStartError: `"SESSION_MODULE_SESSION_START_ERROR"`,
 }
